Experimento2/3001-Socios/requirements.txt
blinker==1.7.0
boto3==1.34.53
botocore==1.34.53
click==8.1.7
Flask==3.0.2
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
marshmallow==3.21.0
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
typing_extensions==4.10.0
urllib3==2.0.7
Werkzeug==3.0.1
